debian: add extra checks when debian/snapd.postrm purge is run

+++ tests/lib/reset.sh
@@ -6,6 +6,10 @@ systemctl stop snapd.service snapd.socket
 
 # purge all state
 sh ${SPREAD_PATH}/debian/snapd.postrm purge
+if [ -d /snap/* ] || [ -d /var/snap/* ]; then
+    echo "postinst purge failed"
+    exit 1
+fi
 mkdir -p /snap /var/snap /var/lib/snapd
 
 rm -f /tmp/ubuntu-core*
